Introduce `TestClient.sendOrQueueSyncResponse`
To ease the path of moving away from matrix-mock-request, introduce a function
in `TestClient` which does the common thing of adding a responder for
`/sync`. Eventually we will swap this out with an alternative implementation
which doesn't depend on matrix-mock-request.

The new function also triggers a `flush` of the sync request, which is
convenient because we invariably follow the existing `.when` with a call to
`flushSync`. So then to actually wait for the sync to happen all we need to do
is await `syncPromise`.