Make some types packages regular dependencies
Due to the fact that we import packages like 'uuid' in files that users of the
library import, these types essentially end up being part of our public library
interface (to TypeScript consumers). However, some of these libraries don't
come with their own types and we import the definitelytyped module, but this is
a dev dependency for us so the types are not available when js-sdk is installed
as a regular dependency. This means type checks on that application will fail.

I think the correct solution here (other than those libraries having their own
types) is to just make the types packages regular dependencies.

This PR does that for the two types packages that element call was failing
because of. I haven't added the rest as they might not all be part of external
facing bits of the library (whatever that is now with js-sdk).