Fix examples/node
Node.js does not understand TypeScript, so the dependency on `file:../..` won't work.
It seems `build:dev` compiles matrix-js-sdk to something in lib/ that Node.js can understand.
So this changes the relative package to `file:../../lib`.
But lib/ is missing a `package.json`, so Node.js doesn't recognize it as a package.
To fix this, I manually add a minimal `package.json`.

Should fix https://github.com/matrix-org/matrix-js-sdk/issues/3145