Fetch events that read receipts point to if we don't have them
As the comment hopefully explains, this is a giant hack to work around
lack of MSC3981 support and hopefully remove some spurious unread dots
on threads. With this code, the unread dots appear briefly and then go
away again once the event in question is fetched which is not fantastic
but better then a perma-stuck dot.